+++ lib/get_it_future_builder.dart
@@ -2,3 +2,5 @@
 library get_it_future_builder;
 
 export 'src/get_it_future_builder.dart';
+export 'src/get_it_future_builder2.dart';
+export 'src/get_it_future_builder3.dart';